pyproject.toml
[project]
name = "groove-on-a-real-train"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "yt-dlp>=2026.3.17",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]

[tool.ruff.lint.isort]
known-first-party = ["groove"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
